+++ hbw/app/javascript/packs/hbw/components/form/select.js.jsx
@@ -115,7 +115,7 @@ modulejs.define('HBWFormSelect',
       isFilled = () => {
         const { value } = this.state;
 
-        return value !== null && value !== undefined && value.length > 0;
+        return value !== null && value !== undefined && `${value}`.length > 0;
       };
 
       setValidationState = () => {